requirements.txt
Django>=4.2
djangorestframework
gunicorn
psycopg2-binary
celery
redis
django-cors-headers
